CRUD de Doadora, Doacao, Agenda e StatusAgenda. StatusAgendaFilter para filtrar staus de agenda.

+++ src/main/java/com/example/amamentai/api/repository/BairroRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.amamentai.api.model.Bairro;
 
-public interface BairroRepository extends JpaRepository<Bairro, Long> {
+public interface BairroRepository extends JpaRepository<Bairro, Integer> {
 
 }
