+++ src/main/java/com/example/amamentai/api/repository/CidadeRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.amamentai.api.model.Cidade;
 
-public interface CidadeRepository extends JpaRepository<Cidade, Long>{
+public interface CidadeRepository extends JpaRepository<Cidade, Integer>{
