+++ src/main/java/com/example/amamentai/api/repository/DoadoraRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.amamentai.api.model.Doadora;
 
-public interface DoadoraRepository extends JpaRepository<Doadora, Long>{
+public interface DoadoraRepository extends JpaRepository<Doadora, Integer>{
 
 }
