+++ src/main/java/com/example/amamentai/api/repository/PessoaRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.amamentai.api.model.Pessoa;
 
-public interface PessoaRepository extends JpaRepository<Pessoa, Long>{
+public interface PessoaRepository extends JpaRepository<Pessoa, Integer>{
